Rework tracing system templates.
The benefit of the new system is that if you feed it a parameter type it
doesn't understand it will no longer compile cleanly but in to an
infinite loop. Also, integral types no longer need coercing to unsigned
long, and neither do pointers.

The downside is the liberal use of dummy template arguments to make the
compiler select the right output function.

Also, add some tests - it's much easier to check compilation and
operation in the test suite. These tests do not currently check the
correct output is made, that's left for the observer.