Add a simple user-side system call library and demo.
The existing system call library is more explicitly labelled as
belonging to the kernel, and a user-side one is created. A simple demo
program in extras/demo_program/ shows how this is used, and replaces the
poorly defined 'testprog' from earlier revisions.

The demo/testprog is also more explicitly added to the build system so
that it is less confusing to get everything working.