Enable RTTI and dynamic_cast
Do this by including libcxxrt from https://github.com/pathscale/libcxxrt
- only the RTTI part of this is used at present. My intention in future
is to include this in the project in a more sensible way than just
copying it from the original, but this will do for now.

Additionally, make a few clarifying comments about how the various
licences apply to the project as a whole.