Add simple file open, read, close syscalls.
Fix a few miscellaneous bugs alongside this.

The main issue at the moment is that it's not possible for a UM program
to learn the size of a file it's attempting to read, and attempting to
read beyond the length of a file causes it to fail. That'll be fixed
shortly.