Add generic PS/2 controller and update IRQ handlers
This is the first stage of work to enable keyboard support.

The IRQ system is changed so that drivers can register handlers for
IRQs, and the system deals with calling all the handlers, rather than
there only being one, generic and useless handler.

The generic PS/2 controller configures all the PS/2 devices in the
system with the devices enabled but scanning and IRQs disabled.

The next phase will be to add a keyboard driver of some sort and plumb
it in.