Begin a keyboard driver and handle IRQs better
It's very difficult to handle IRQs entirely in the context of the
interrupt itself, since it makes locking very hard. As such, introduce a
"slow path" which is called after the main IRQ handler in the context of
a kernel thread.

Begin a keyboard driver that simply uses messages to send the scancodes
to whichever process it is told to.

This mechanism will hopefully be built on to enable Azalea's first
terminal.