Move to new folder layout
As the project develops, I want to be able to write various user-mode
tools. This will necessitate a user-facing library / SDK-type-thing. It
makes sense to include these in the same GIT repository, so it also
makes sense to adjust the folder tree to make the split between
components more obvious.

Kernel-specific parts are now in kernel/, user mode programs are in
user/ and the necessary libraries will go in user/libs. Components
introduced from outside still live in external/.

This commit includes no changes other than those needed to make the
project compile.