Separate out user mode visible interfaces
Move any interface that is shared with user-mode into a dedicated
folder. For example, error codes and structures passed across the
syscall boundary. Do this for two reasons:

1 - It's much clearer whether a given interface item is used in user-
mode code as well.
2 - It should make it easier to construct a SDK at some point in the
future.

Also, restructure the build script. The idea here is that it'll one day
be possible to request the building of only specific parts of the
project.