Add process and thread control system calls
User mode processes can now start and control threads and processes. The
demo program is updated to demonstrate this. This is not flawless - for
example, handles are not currently released by exiting threads.

In addition, fix a couple of bugs that should have been obvious, but
didn't manifest in QEMU for whatever reason:

1 - There was no way the AP trampoline would work on a real machine,
since it was linked in an arbitrary position. Goodness knows how QEMU
executed it! It is now re-linked against a sensible entry vector.

2 - Having more that one user-mode thread corrupted the stack on each
task switch because we weren't paying enough attention to the stack.
This whole area of code is given a bit of a tidy up, alongside switching
to using the interrupt stack table mechanism to ensure a more consistent
way of needing to deal with the stack across a task switch.