Introduce a Windows build of the unit test program
Technical issues still prevent being able to build the entire kernel on
Windows, but this will allow some work to be done without requiring a
Linux machine or compilation toolchain. It also opens up the option of
using the Visual Studio debugger to work on parts of the kernel code
covered by the unit test program.