Introduce a memory based filesystem and switch to smart pointers.
The memory filesystem allows for creation, reading and writing of files.
Deleting files and manipulation of folders will be added in a later
revision.

Smart pointers are now used to manage object lifetimes instead of the
reference counting system used before - it is easier to get right!

Because of a change to the standard location of stdout, revision b5897d9
or later of Azalea libc is now required.