Stop requiring disk images.
QEMU can boot a multiboot kernel image directly, and also create a FAT
partition directly from a folder on the host machine disk. That is more
than enough to provide a demo virtual machine.

This has the significant advantage of not needing to fiddle with disk
images during a normal code-build-test loop, and makes it easier to
develop using WSL.