Add code for C-library time to the shell
An upcoming version of the Azalea C Library will add support for time.h
functions. This will require the addition of timezone data to the system
image, so a note is added to the build script to remind the user to do
this.

It's not done automatically because of future plans for big changes to
the layout of the code and build system.