Correct futexes and add 'online tests'
Futexes in Azalea are intended to be based on virtual addresses, like in
Fuchsia. Also the check to see if the value was still as expected was
incorrect.

'Online tests' are tests that can't be carried out in the old style unit
tests very easily - they are much better done with a running kernel. At
the moment there are very few tests, the intent is to expand them over
time.