Bring the main build script back to this project.
Having separate build scripts seemed to be a good idea when the kernel
was gaining dependencies on external projects. The aim was to make the
build script handle those dependencies, however it never really worked
out that way and was just a bit of a nuisance.

Instead, it seems more sensible to codify dependencies within the main
tree, probably by using git submodules.